UottawaHack/requirements.txt
google-generativeai>=0.3.0
PyPDF2>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pymongo>=4.6.0
python-dotenv>=1.0.0
reportlab>=4.0.0
